requirements.txt
ccxt=4.4.42
pandas=2.2.3
numpy=2.2.1
python-telegram-bot=21.9
